+++ FortigateManager/functions/Connect-FM.ps1
@@ -15,10 +15,20 @@
 	.PARAMETER ADOM
 	The default ADOM for the requests.
 
+	.PARAMETER OldConnection
+	An old connection to be revived. This can be obtained e.g. by Export-Clixml/Import-Clixml.
+
     .PARAMETER SkipCheck
     Array of checks which should be skipped while using Invoke-WebRequest.
     Possible Values 'CertificateCheck', 'HttpErrorCheck', 'HeaderValidation'.
     If neccessary by default for the connection set $connection.SkipCheck
+
+	.PARAMETER Type
+	To which type of endpoint server should the connection be established?
+	Manager -  FortiManager
+	Analyzer - Forti Analyzer
+	Connections of type 'manager' (default) are used for [Verb]-FM[Noun] commands,
+	type 'Analyzer' is needed for the [Verb]-FMA[Noun] commands
 
 	.PARAMETER EnableException
 	Should Exceptions been thrown?
@@ -27,6 +37,13 @@
 	$connection=Connect-FM -Url $url -Credential $cred
 
 	Connect directly with a Credential-Object
+	.EXAMPLE
+	$connection=Connect-FM -Url $url -Credential $cred
+	$connection=Export-Clixml -Path ".\connection.xml"
+	$importedConnection=Import-Clixml -Path ".\connection.xml"
+	$secondConnection=Connect-FM -OldConnection $importedConnection
+
+	Connect with the information from a serialized object
 
 	.NOTES
 	#>
@@ -42,20 +59,41 @@
 		[string]$ADOM,
 		[parameter(mandatory = $true, ParameterSetName = "credential")]
 		[pscredential]$Credential,
+		[parameter(mandatory = $true, ParameterSetName = "oldConnection")]
+		$OldConnection,
 		[ValidateSet('CertificateCheck', 'HttpErrorCheck', 'HeaderValidation')]
 		[String[]]$SkipCheck,
+		[parameter(mandatory = $false, ParameterSetName = "credential")]
+		[ValidateSet("Manager", "Analyzer")]
+		[string]$Type = "Manager",
 		[bool]$EnableException = $true
 	)
 	begin {
+		if ($OldConnection) {
+			Write-PSFMessage "Getting parameters from existing (mistyped) Connection object"
+			$Url = $OldConnection.ServerRoot
+			$Credential = $OldConnection.credential
+			if ($OldConnection.SkipCheck) {
+				$connection.SkipCheck
+			}
+			$additionalParams = $OldConnection.forti
+			if ($OldConnection.forti.defaultADOM) {
+				$ADOM = $OldConnection.forti.defaultADOM
+			}
+		}
+		else {
+			$additionalParams = @{
+				requestId       = 1
+				session         = $null
+				EnableException = $EnableException
+				ConnectionType  = $Type
+			}
+		}
 	}
 	end {
 		$connection = Get-ARAHConnection -Url $Url -APISubPath ""
-		if ($SkipCheck) { $connection.SkipCheck = $SkipCheck}
-		Add-Member -InputObject $connection -MemberType NoteProperty -Name "forti" -Value @{
-			requestId = 1
-			session   = $null
-			EnableException=$EnableException
-		}
+		if ($SkipCheck) { $connection.SkipCheck = $SkipCheck }
+		Add-Member -InputObject $connection -MemberType NoteProperty -Name "forti" -Value $additionalParams
 		$connection.credential = $Credential
 		$connection.ContentType = "application/json;charset=UTF-8"
 		$connection.authenticatedUser = $Credential.UserName
@@ -96,10 +134,16 @@
 				$this.forti.session = $result.session
 			}
 		}
-		$connection.Refresh()
+		switch ($PsCmdlet.ParameterSetName) {
+			'credential' {
+				$connection.Refresh()
+			}
+			'oldConnection' {}
+		}
 		if ($connection.forti.session) {
 			Write-PSFMessage -string "Connect-FM.Connected"
-			Set-PSFConfig -Module 'FortigateManager' -Name 'LastConnection' -Value $connection -Description "Last known Connection" -AllowDelete
+			Write-PSFMessage "Saving last connection as Config LastConnection.$($connection.forti.ConnectionType)"
+			Set-PSFConfig -Module 'FortigateManager' -Name "LastConnection.$($connection.forti.ConnectionType)" -Value $connection -Description "Last known Connection" -AllowDelete
 			return $connection
 		}
 		Write-PSFMessage -string "Connect-FM.NotConnected" -Level Warning