+++ FortigateManager/functions/Rename-FMAddress.ps1
@@ -77,7 +77,7 @@
             data = @{name = $Mapping.$oldName }
         }
     }
-    # Write-PSFMessage -Level Host "`$apiCallParameter=$($apiCallParameter|ConvertTo-PSFHashtable -Exclude connection |ConvertTo-Json -Depth 4)"
+    # Write-PSFMessage -Level Host "`$apiCallParameter=$($apiCallParameter|ConvertTo-PSFHashtable -Exclude connection | ConvertTo-Json -WarningAction SilentlyContinue -Depth 4)"
     $result = Invoke-FMAPI @apiCallParameter
     if (-not $EnableException) {
         return $result